SR-2850 Post and put payloads not recognized.
Fixes a bug caused by the fact that the `fetch` function did
not automatically set the `Content-Type` HTTP header to
`application/json` when sending a payload in this format.

This change only bumps patch segment of the version because is
a bugfix. However, for those how may have managed to make the
code in 1.2.0 work this update **does include breaking changes**.