Finishes implementing the tests.
Beware the fact that, in `document_tests_helper.js` a placeholder
`_ => true` function has been passed to the `query` nock mocks
when `auth_token` was the only passed GET parameter. This is the
result of me desisting to trying to understand how on earth
the GET query matching works for Nock. It is no big deal, though,
since we are not testing authentication here.